requirements.txt
beautifulsoup4
html2text
python-dotenv
requests
